+++ launcher/internal/process/job_other.go
@@ -1,0 +1,9 @@
+//go:build !windows
+
+package process
+
+import "os/exec"
+
+func assignToJobObject(cmd *exec.Cmd) error {
+	return nil
+}